backend/pyproject.toml
[project]
name = ""
version = "0.0.1"
requires-python = "==3.11.4"
dependencies = [
    "click==8.0.4",
    "django==5.0.2",
    "psycopg2-binary==2.9.10",
    "requests==2.32.3",
    "twilio==9.3.6",
    "channels==4.1.0",
    "channels-redis==4.2.0",
    "phonenumbers==8.13.49",
    "celery==5.2.7",
    "redis==5.2.0",
    "sentry-sdk==2.18.0",
    "dj-database-url==2.3.0",
    "djangorestframework==3.15.2",
    "unidecode==1.3.8",
    "coreapi==2.3.3",
    "django-filter==24.3",
    "shortener==0.2.1",
    "django-labs-accounts==0.8.0",
    "beautifulsoup4==4.12.3",
    "ipython==8.29.0",
    "django-auto-prefetching==0.2.12",
    "django-cors-headers==4.6.0",
    "django-runtime-options==0.1.3",
    "pyyaml==6.0.2",
    "uritemplate==4.1.1",
    "uwsgi==2.0.28",
    "uvloop==0.21.0",
    "gunicorn==23.0.0",
    "uvicorn[standard]==0.32.0",
    "httptools==0.6.4",
    "ddt==1.7.2",
    "lark-parser==0.12.0",
    "tqdm==4.67.0",
    "boto3==1.35.57",
    "nose==1.3.7",
    "importlib-metadata==8.5.0",
    "django-redis==5.4.0",
    "jsonref==1.1.0",
    "tblib==3.0.0",
    "django-extensions==3.2.3",
    "numpy==2.1.3",
    "scikit-learn==1.5.2",
    "pandas==2.2.3",
    "python-dateutil==2.9.0.post0",
    "docutils==0.21.2",
    "ics==0.7.2",
    "drf-nested-routers==0.94.1",
    "google-api-python-client==2.151.0",
    "asyncio==3.4.3",
    "aiohttp==3.10.10",
]

[tool.black]
line-length = 100

[dependency-groups]
dev = [
    "coverage==7.6.4",
    "unittest-xml-reporting==3.2.0",
    "flake8==7.1.1",
    "flake8-isort==6.1.1",
    "isort==5.13.2",
    "flake8-quotes==3.4.0",
    "pytest==8.3.3",
    "pytest-django==4.9.0",
    "django-debug-toolbar==4.4.6",
    "black==22.3.0",
]

[tool.uv]
package = false

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
